+++ src/commands/dev/generate/library.ts
@@ -17,11 +17,11 @@ const messages = Messages.load('@salesforce/plugin-dev', 'dev.generate.library',
 
 export default class GenerateLibrary extends SfCommand<void> {
   public static enableJsonFlag = false;
-  public static summary = messages.getMessage('summary');
-  public static description = messages.getMessage('description');
-  public static examples = messages.getMessages('examples');
+  public static readonly summary = messages.getMessage('summary');
+  public static readonly description = messages.getMessage('description');
+  public static readonly examples = messages.getMessages('examples');
 
-  public static flags = {};
+  public static readonly flags = {};
 
   public async run(): Promise<void> {
     await generate('library', { force: true });