+++ src/commands/dev/generate/plugin.ts
@@ -13,11 +13,12 @@ const messages = Messages.load('@salesforce/plugin-dev', 'dev.generate.plugin', 
 
 export default class GeneratePlugin extends SfCommand<void> {
   public static enableJsonFlag = false;
-  public static summary = messages.getMessage('summary');
-  public static description = messages.getMessage('description');
-  public static examples = messages.getMessages('examples');
-
-  public static flags = {};
+  public static readonly summary = messages.getMessage('summary');
+  public static readonly description = messages.getMessage('description');
+  public static readonly examples = messages.getMessages('examples');
+  public static aliases = ['plugins:generate'];
+  public static readonly deprecateAliases = true;
+  public static readonly flags = {};
 
   public async run(): Promise<void> {
     await generate('plugin', { force: true });