+++ test/commands/dev/convert/samples/messages/basic-js.js
@@ -1,0 +1,21 @@
+module.exports = {
+  "string": "foo",
+  "stringTrimmed": "foo\n\n",
+  "multiline": "foo\n\nbar",
+  "array": [
+    "boof",
+    "bark"
+  ],
+  "object": {
+    "bink": "bunk"
+  },
+  "nested": {
+    "chicken": {
+      "egg": "chicken",
+      "chicken": "feather"
+    },
+    "dog": {
+      "blanket": "dog"
+    }
+  }
+};